backend/requirements.txt
# FastAPI and ASGI server
fastapi==0.104.1
uvicorn[standard]==0.24.0
gunicorn==21.2.0

# Database and ORM
sqlalchemy==2.0.23
alembic==1.12.1
asyncpg==0.29.0
psycopg2-binary==2.9.9

# Redis and caching
redis==5.0.1
hiredis==2.2.3

# Authentication and security
python-jose[cryptography]==3.3.0
passlib[bcrypt]==1.7.4
python-multipart==0.0.6
cryptography==41.0.7

# HTTP client and WebSockets
httpx==0.25.2
websockets==12.0

# Background tasks
celery==5.3.4
flower==2.0.1

# File processing and AI/ML
PyMuPDF==1.23.8
python-pptx==0.6.23
openpyxl==3.1.2
python-docx==1.1.0
Pillow==10.1.0

# AI/ML libraries
transformers==4.35.2
torch==2.1.1
spacy==3.7.2
scikit-learn==1.3.2
pandas==2.1.3
numpy==1.24.4
scipy==1.11.4

# Document processing
sentence-transformers==2.2.2
nltk==3.8.1

# Cloud storage
boto3==1.34.0
minio==7.2.0

# Search and analytics
elasticsearch==8.11.0

# Monitoring and logging
prometheus-client==0.19.0
structlog==23.2.0

# Development and testing
pytest==7.4.3
pytest-asyncio==0.21.1
pytest-cov==4.1.0
black==23.11.0
isort==5.12.0
flake8==6.1.0
mypy==1.7.1

# Environment and configuration
python-dotenv==1.0.0
pydantic==2.5.0
pydantic-settings==2.1.0

# Email and notifications
emails==0.6.0
jinja2==3.1.2

# Utilities
python-slugify==8.0.1
python-dateutil==2.8.2
pytz==2023.3
